Data Stores Implemented

+++ src/data/Cleaning.java
@@ -5,7 +5,7 @@ public class Cleaning extends Treatment{
     private static Cleaning instance;
 
     //Parameterized Constructor to initialize the attributes upon instantiation:
-    private Cleaning(String treatmentType, float fee, short timeInMinutes) {
+    private Cleaning(String treatmentType, double fee, short timeInMinutes) {
 
         this.treatmentType = treatmentType;
         this.fee = fee;
@@ -13,7 +13,7 @@ public class Cleaning extends Treatment{
 
     }
 
-     protected static Cleaning getCleaning(String treatmentType, float fee, short timeInMinutes){
+     protected static Cleaning getCleaning(String treatmentType, double fee, short timeInMinutes){
 
         if(instance == null){
             instance = new Cleaning(treatmentType, fee, timeInMinutes);
