+++ src/data/NerveFilling.java
@@ -5,7 +5,7 @@ public class NerveFilling extends Treatment{
     private static NerveFilling instance;
 
     //Parameterized Constructor to initialize the attributes upon instantiation:
-    private NerveFilling(String treatmentType, float fee, short timeInMinutes) {
+    private NerveFilling(String treatmentType, double fee, short timeInMinutes) {
 
         this.treatmentType = treatmentType;
         this.fee = fee;
@@ -13,7 +13,7 @@ public class NerveFilling extends Treatment{
 
     }
 
-    protected static NerveFilling getNerveFilling(String treatmentType, float fee, short timeInMinutes){
+    protected static NerveFilling getNerveFilling(String treatmentType, double fee, short timeInMinutes){
 
         if(instance == null){
             instance = new NerveFilling(treatmentType, fee, timeInMinutes);
